changes file structure -- ITS CLEAN
