+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import './App.css';
 // import ReactGa from 'react-ga';
-import Navbar from './components/Navbar.js';
-import Footer from './components/Footer.js';
-import Routes from './components/Routes.js';
+import Navbar from './components/Nav/Navbar.js';
+import Footer from './components/Landing/Footer.js';
+import Routes from './components/Routes/Routes.js';
 import MockEvents from './components/MockEvents';
 
 function App() {
