+++ src/components/Authorized/Login.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
-import { ModalStyle, getModalStyle } from '../styles/ModalFormStyles.js';
+import { ModalStyle, getModalStyle } from '../../styles/ModalFormStyles.js';
 import { withRouter } from 'react-router-dom';
 import Modal from '@material-ui/core/Modal';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
-import { axiosWithAuth } from '../utils/axiosWithAuth.js';
+import { axiosWithAuth } from '../../utils/axiosWithAuth.js';
 
 const useStyles = ModalStyle;
