+++ src/components/Nav/Event-Page.js
@@ -1,0 +1,101 @@
+import React, { useState, useEffect } from 'react';
+import { withRouter } from 'react-router-dom';
+import axios from 'axios';
+
+const EventPage = (props) => {
+  const [string, setString] = useState('');
+  const [result, setRes] = useState([]);
+  const [time, setTime] = useState({
+    startTime: '',
+    endTime: '',
+    startDate: '',
+    endDate: '',
+  });
+
+  console.log(props)
+
+  const eventID = Number(props.match.params.id);
+
+  useEffect(() => {
+    axios
+      .get(`https://phoenix-be-staging.herokuapp.com/api/calendar/`)
+      .then(async (res) => {
+        let results = await res.data.filter((item) => {
+          return item.summary
+            .toLowerCase()
+            .includes(string.toLocaleLowerCase());
+        });
+
+        let event = await results.filter((item) => {
+          return item.event_id === eventID;
+        })[0];
+
+        setRes(event);
+
+        //convert start time
+        let startHour = Number(event.start_time.substr(0, 2));
+        let startMin = event.start_time.substr(3, 2);
+
+        let startT;
+        if (startHour < 12) {
+          startT = `${startHour}:${startMin} AM`;
+        }
+        if (startHour > 12) {
+          startT = `${(startHour -= 12)}:${startMin} PM`;
+        }
+        if (startHour === 12) {
+          startT = `${startHour}:${startMin} PM`;
+        }
+
+        // convert end time
+        let endHour = Number(event.end_time.substr(0, 2));
+        let endMin = event.end_time.substr(3, 2);
+
+        let endT;
+        if (endHour < 12) {
+          endT = `${endHour}:${endMin} AM`;
+        }
+        if (endHour > 12) {
+          endT = `${(endHour -= 12)}:${endMin} PM`;
+        }
+        if (endHour === 12) {
+          endT = `${endHour}:${endMin} PM`;
+        }
+
+        // convert date
+        let start = event.start_date;
+        let startD = start.substr(0, 10);
+        let end = event.end_date;
+        let endD = end.substr(0, 10);
+
+        setTime({
+          startTime: startT,
+          endTime: endT,
+          startDate: startD,
+          endDate: endD,
+        });
+      })
+      .then((_) => setString(''))
+      .catch((err) => console.log('Problem retrieving events', 'Error: ', err));
+
+    if (result.start_time > '12:00:00') {
+      let startO = result.start_time - 12 + 'pm';
+      setTime({ startTime: startO });
+    }
+  }, [eventID]);
+
+  return (
+    <>
+      <h2>{result.summary}</h2>
+      <p>{result.description}</p>
+      <p>Location: {result.location}</p>
+      <p>
+        Starts: {time.startDate} {time.startTime}
+      </p>
+      <p>
+        Ends: {time.endDate} {time.endTime}
+      </p>
+    </>
+  );
+};
+export default withRouter(EventPage);
