+++ src/components/RegisteredDash/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import Calendar from './Calendar/Calen.js';
-import EventList from './RegisteredEvents/EventList.js';
-import DashboardStyles from '../styles/dashboardstyle.js';
+import Calendar from '../Calendar/Calen.js';
+import EventList from '../RegisteredEvents/EventList.js';
+import DashboardStyles from '../../styles/dashboardstyle.js';
 import Paper from '@material-ui/core'
 
 const Dashboard = () => {
